feat(ui): match app.quickgig.ph styling; fix Next metadata build error by moving metadata to server component; add color tokens
