+++ src/components/ui/Button.tsx
@@ -12,13 +12,13 @@ interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
 
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   ({ className, variant = 'primary', size = 'md', children, loading, disabled, ...props }, ref) => {
-    const baseClasses = 'qg-btn inline-flex items-center justify-center gap-2 font-heading font-semibold transition-all duration-qg-fast focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed';
-    
+    const baseClasses = 'inline-flex items-center justify-center gap-2 font-heading font-semibold transition-all duration-qg-fast focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed';
+
     const variants = {
-      primary: 'qg-btn-primary focus:ring-qg-primary',
-      secondary: 'qg-btn-secondary focus:ring-qg-accent',
-      outline: 'qg-btn-outline focus:ring-qg-primary',
-      ghost: 'bg-transparent text-qg-primary hover:bg-qg-primary/10 focus:ring-qg-primary',
+      primary: 'bg-primary text-white hover:bg-primary/90 focus:ring-primary',
+      secondary: 'bg-secondary text-fg hover:bg-secondary/90 focus:ring-secondary',
+      outline: 'border border-primary text-primary hover:bg-primary hover:text-white focus:ring-primary',
+      ghost: 'bg-transparent text-primary hover:bg-primary/10 focus:ring-primary',
     };
     
     const sizes = {
