+++ tailwind.config.ts
@@ -9,8 +9,12 @@ const config: Config = {
   theme: {
     extend: {
       colors: {
-        background: "var(--background)",
-        foreground: "var(--foreground)",
+        primary: 'var(--color-primary)',
+        secondary: 'var(--color-secondary)',
+        bg: 'var(--bg)',
+        fg: 'var(--fg)',
+        background: 'var(--bg)',
+        foreground: 'var(--fg)',
         // QuickGig Brand Colors
         qg: {
           primary: {
